% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_logging.R
\name{configure_logging}
\alias{configure_logging}
\title{Configure logging options}
\usage{
configure_logging(threshold_level = "DEBUG", output_file = "")
}
\arguments{
\item{threshold_level}{Defines threshold level for printing logs.
Should be one of 'DEBUG', 'INFO', 'WARNING' or 'ERROR'. Default to 'DEBUG'.
It defines which logs should be printed. It prints logs which have level higher or equal to threshold_level.
The hierarchy is 'DEBUG' < 'INFO' < 'WARNING' < 'ERROR'.}

\item{output_file}{Either a path to a file or a connection. With default settings "" logs are printed to console.}
}
\description{
Configure logging options
}
\details{
\code{configure_logging} sets global behaviour for logging and acts as a filter of how detailed logs should appear in the output.
}
\examples{

# set logging threshold level to INFO
configure_logging(threshold_level = 'INFO')

# set logging threshold level to WARNING and
# set writing all logs to 'log.log' file
configure_logging(threshold_level = 'WARNING', output_file = 'log.log')

}
\seealso{
\code{\link{create_logger}}
}

