% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureLogging.R
\name{configureLogging}
\alias{configureLogging}
\title{Configure logging options}
\usage{
configureLogging(level = "DEBUG", file = "")
}
\arguments{
\item{level}{A logging level. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'. Default is 'DEBUG'}

\item{file}{A connection, or a character string naming the file to print to.}
}
\description{
Configure logging options
}
\examples{

# set logging level to INFO
configureLogging(level = 'INFO')

}

