% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextManager.R
\name{ContextManager}
\alias{ContextManager}
\title{Create a Context Manager}
\usage{
ContextManager(start = function() { }, end = function() { })
}
\arguments{
\item{start}{A function to be executed on start}

\item{end}{A function to be executed on exit}
}
\value{
A function
}
\description{
Create a Context Manager
}

