% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logger.R
\name{create_logger}
\alias{create_logger}
\title{Create a logger}
\usage{
create_logger(name, level, file = "")
}
\arguments{
\item{name}{A logger name}

\item{level}{A logging level. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'.}

\item{file}{A connection, or a character string naming the file to print to.}
}
\description{
Create a logger
}
\examples{
# create an object of class Logger with name 'clean_data' and
# default logging level INFO
logger <- create_logger(name = 'clean_data', level = 'INFO')

# create log in context of logger
logger$log_message('Some text')

# output:
# 2016-09-21 10:59:22.0 - [clean_data] - INFO - [Some text]


# create log with changed level to DEBUG in context of logger
logger$log_message('Some text', level = 'DEBUG')

# output:
# 2016-09-21 11:00:04.1 - [clean_data] - DEBUG - [Some text]


# create logtime in context of logger
logger$log_time('Some text') \%<\% {
    Sys.sleep(1)
}

# output:
# 2016-09-21 11:01:25.6 - [clean_data] - INFO - [Start] - [Some text]
# 2016-09-21 11:01:26.7 - [clean_data] - INFO - [End] - [Some text] - 
#   [Done in 1 sec. (0 min.)]


# create logtime with changed level to DEBUG in context of logger
logger$log_time('Some text', level = 'DEBUG') \%<\% {
    Sys.sleep(1)
}

# output:
# 2016-09-21 11:03:43.7 - [clean_data] - DEBUG - [Start] - [Some text]
# 2016-09-21 11:03:44.8 - [clean_data] - DEBUG - [End] - [Some text] - 
#   [Done in 1 sec. (0 min.)]


}
\seealso{
\code{\link{configure_logging}}
}

