% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_logging.R
\name{get_configs}
\alias{get_configs}
\title{Get logging configs}
\usage{
get_configs()
}
\description{
Get logging configs which were set in \code{configure_logging}
}
\examples{

# get logging configs
get_configs()
# Threshold Logging Level:  DEBUG
# Output Logging File:  NA (Print to console) 

}
\seealso{
\code{\link{configure_logging}}
}

