% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_logging.R
\name{get_logging_configs}
\alias{get_logging_configs}
\title{Get logging configs}
\usage{
get_logging_configs()
}
\value{
A list with 'logging_level' and 'file'
}
\description{
Get logging configs which were set in \code{configure_logging}
}
\examples{

get_logging_configs()
# $logging_level
# [1] "DEBUG"
# 
# $file
# [1] ""


}
\seealso{
\code{\link{configure_logging}}
}

