% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log}
\alias{log}
\title{Print Log message}
\usage{
log(message, level = 'DEBUG')
}
\arguments{
\item{message}{A string}

\item{level}{A logging level. One of 'INFO' or 'DEBUG'. Default is DEBUG.}
}
\value{
Prints a message
}
\description{
Create log message with specific level
}
\examples{

# create a log with message 'Some text'
log('Some text')  # default level is DEBUG

# output:
# 2016-09-16 16:17:20.9 - DEBUG - [Some text]


# create the same log with changed level to INFO
log('Some text', level = 'INFO')

# output:
# 2016-09-16 16:18:00.5 - INFO - [Some text]

}

