% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{log_message}
\alias{log_message}
\title{Print Log message}
\usage{
log_message(message, level = 'DEBUG')
}
\arguments{
\item{message}{A string}

\item{level}{A logging level. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'. Default is 'DEBUG'.}
}
\value{
Prints a message
}
\description{
Create log message with specific level
}
\examples{

# create a log with message 'Some text'
log_message('Some text')  # default level is DEBUG

# output:
# 2016-09-21 10:55:18.8 - DEBUG - [Some text]


# create the same log with changed level to INFO
log_message('Some text', level = 'INFO')

# output:
# 2016-09-21 10:55:38.0 - INFO - [Some text]

}
\seealso{
\code{\link{log_time}}
}

