% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_time.R
\name{log_time}
\alias{log_time}
\title{Track time of code execution}
\usage{
log_time(message, level = 'DEBUG') \%<\% \{
   expression
\}
}
\arguments{
\item{message}{A string describing context of code}

\item{level}{A logging level. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'. Default is DEBUG.}
}
\value{
Prints 2 log messages with start, end of code execution and time of code execution
}
\description{
Time your code in easy, efficient and nice looking way
}
\examples{

# create a logtime with message 'Some text';
# default level is DEBUG
log_time('Some text') \%<\% {
   Sys.sleep(1)
}

# output:
# 2016-09-21 10:54:31.8 - DEBUG - [Start] - [Some text]
# 2016-09-21 10:54:32.9 - DEBUG - [End] - [Some text] - [Done in 1 sec. (0 min.)]


# create the same logtime with changed level to INFO
log_time('Some text', level = 'INFO') \%<\% {
   Sys.sleep(1)
}

# output:
# 2016-09-21 10:53:20.6 - INFO - [Start] - [Some text]
# 2016-09-21 10:53:21.6 - INFO - [End] - [Some text] - [Done in 1 sec. (0 min.)]

}
\seealso{
\code{\link{log_message}}
}

