% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logtime.R
\name{Logtime}
\alias{Logtime}
\title{Track time of code execution}
\usage{
Logtime(message) \%<\% \{
   expression
\}
}
\arguments{
\item{message}{A string describing context of code}
}
\value{
Prints 2 log messages with start, end of code execution and prints time of code execution
}
\description{
Time your code in easy, efficient and nice looking way
}
\examples{
Logtime('Generate random numbers') \%<\% {
     rnorm(10000) 
}

# output:
#


}

