% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logtime.R
\name{OnEnd}
\alias{OnEnd}
\title{Print ending log with message and exec time after execution of 
expresion in context of Logtime}
\usage{
OnEnd(msg)
}
\arguments{
\item{msg}{A message to print in log}
}
\description{
Print ending log with message and exec time after execution of 
expresion in context of Logtime
}

