% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logtime.R
\name{OnStart}
\alias{OnStart}
\title{Print starting log with message before execution of 
expresion in context of Logtime}
\usage{
OnStart(msg)
}
\arguments{
\item{msg}{A message to print in log}
}
\description{
Print starting log with message before execution of 
expresion in context of Logtime
}

