% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_operator.R
\name{\%<\%}
\alias{\%<\%}
\title{Pipe operator for log_time}
\usage{
lhs \%<\% rhs
}
\arguments{
\item{lhs}{A function call which first argument is expression}

\item{rhs}{An expression}
}
\description{
Pipe operator for log_time
}

