% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_logging.R
\name{configure_logging}
\alias{configure_logging}
\title{Configure logging options}
\usage{
configure_logging(threshold_level = "DEBUG", output_file = "")
}
\arguments{
\item{threshold_level}{A threshold logging level for printing logs. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'. Default is 'DEBUG'.}

\item{output_file}{A connection, or a character string naming the file to print to. By default all logs are being printed to console.}
}
\description{
Configure logging options
}
\examples{

# set logging threshold level to INFO
configure_logging(threshold_level = 'INFO')

}

