% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{log_message}
\alias{log_message}
\title{Print Log message}
\usage{
log_message(message, level = logger_level)
}
\arguments{
\item{message}{A string}

\item{level}{A logging level. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'. Default is 'DEBUG'.}
}
\value{
Prints log to console or file. This can be controlled with \code{configure_logging} or \code{create_logger} if \code{log_message} is used inside logger context.
}
\description{
Print Log message
}
\examples{

# create a log with message 'Some text'
log_message('Some text')  # default level is DEBUG
# output:
# 2016-09-21 10:55:18 - DEBUG - [Some text]


# create the same log with changed level to INFO
log_message('Some text', level = 'INFO')
# output:
# 2016-09-21 10:55:38 - INFO - [Some text]

}
\seealso{
\code{\link{log_time}}
}

