% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logtime.R
\name{logtime}
\alias{logtime}
\title{Track time of code execution}
\usage{
logtime(message, level = 'DEBUG') \%<\% \{
   expression
\}
}
\arguments{
\item{message}{A string describing context of code}

\item{level}{A logging level. One of 'DEBUG', 'INFO', 'WARNING', 'ERROR'. Default is DEBUG.}
}
\value{
Prints 2 log messages with start, end of code execution and time of code execution
}
\description{
Time your code in easy, efficient and nice looking way
}
\examples{

# create a logtime with message 'Some text';
# default level is DEBUG
logtime('Some text') \%<\% {
   Sys.sleep(1)
}

# output:
# 2016-09-16 16:19:29.8 - DEBUG - [Start] - [Some text]
# 2016-09-16 16:19:30.8 - DEBUG - [End] -
#  [Some text] - [Done by 1 sec. (0.02 min.)]


# create the same logtime with changed level to INFO
logtime('Some text', level = 'INFO') \%<\% {
   Sys.sleep(1)
}

# output:
# 2016-09-16 16:24:15.7 - INFO - [Start] - [Some message]
# 2016-09-16 16:24:16.7 - INFO - [End] -
#  [Some message] - [Done by 1 sec. (0.02 min.)]

}

