% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOperator.R
\name{\%<\%}
\alias{\%<\%}
\title{Pipe operator for Context Manager}
\usage{
lhs \%<\% rhs
}
\arguments{
\item{lhs}{A function call which first argument is expression}

\item{rhs}{An expression}
}
\description{
Pipe operator for Context Manager
}

